Extract character replacing CharSequence to a separate file
